make client load more posts as you reach the end of the page

+++ webapp/src/components/interactions.jsx
@@ -1,21 +1,16 @@
 import axios from "axios";
+import { useEffect, useState } from "react";
 import { useFetcher } from "react-router-dom";
 
 export async function likeAction({ request, params }) {
   try {
-    let response = await axios.post("/posts/like?id=" + params.postId);
-    if (response.status === 200) {
-      return response.data;
+    let formData = await request.formData();
+    let response = null;
+    if (formData.get("like") === "true") {
+      response = await axios.post("/posts/like?id=" + params.postId);
+    } else {
+      response = await axios.delete("/posts/like?id=" + params.postId);
     }
-  } catch (error) {
-    console.log(error);
-  }
-  return null;
-}
-
-export async function unlikeAction({ request, params }) {
-  try {
-    let response = await axios.delete("/posts/like?id=" + params.postId);
     if (response.status === 200) {
       return response.data;
     }
@@ -32,20 +27,28 @@ export default function Interactions({
   replyCount,
 }) {
   const fetcher = useFetcher();
+  const [liked, setLiked] = useState(likedByUser);
+  const [newLikeCount, setNewLikeCount] = useState(likeCount);
+
+  useEffect(() => {
+    if (fetcher.formData) {
+      let liked = fetcher.formData.get("like") === "true";
+      setLiked(liked);
+      setNewLikeCount(newLikeCount => liked ? newLikeCount + 1 : newLikeCount - 1);
+    }
+  }, [fetcher.formData]);
 
   return (
     <div className="hstack gap-2 justify-content-evenly text-center">
-      <fetcher.Form
-        method="post"
-        action={likedByUser ? `/post/${postId}/unlike` : `/post/${postId}/like`}
-      >
-        <button className="btn link-danger btn-md">
-          {likedByUser ? (
-            <i className="bi bi-heart-fill"></i>
-          ) : (
-            <i className="bi bi-heart"></i>
-          )}{" "}
-          {likeCount} curtidas
+      <fetcher.Form method="post" action={`/post/${postId}/like`}>
+        <button
+          className="btn link-danger btn-md"
+          name="like"
+          value={liked ? "false" : "true"}
+        >
+          <i className={liked ? "bi bi-heart-fill" : "bi bi-heart"} />{" "}
+          {newLikeCount}
+          {newLikeCount === 1 ? " curtida" : " curtidas"}
         </button>
       </fetcher.Form>
       <button className="btn link-primary btn-md">
